requirements.txt
Flask>=3.0
psycopg2-binary>=2.9
python-dotenv>=1.0
openai>=1.0.0